SearchApi and Store-Vuex
Development of the searchApi component, the inputs and the button, they were also stylized and functionality for data collection was added. In the store, the initial states were defined, two functions were added in mutations to add the data from the searchApi component to the global state of the app.