MovieScreen
The addMovieView component was developed, it consists of inputs, references and the call to the store to add the new data, in turn while editing the new local movie it appears in the lower part. Added MovieScreen component that displays more card information when clicking on (More...). The functionality and styles were implemented, in addition the selected data is stored in local storage to be consumed later. Added to MovieScreen router, imported into axios store. In addition, the form was configured for the SearchApi component so that when submitting, it would not do the reloading.